pyproject.toml
[project]
name = "ai-with-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "networkx>=3.4.2",
    "openai>=2.3.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
